Extract cache deletion side-effects on expired cache from the ´load´ method to the ´validateCache´ method. Extract common methods into a new file
