+++ EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -49,7 +49,6 @@ public final class LocalFeedLoader {
             case let .found(images, timestamp) where self.validate(timestamp):
                 completion(.success(images.toModels()))
             case .found:
-                self.store.deleteCachedFeed { _ in }
                 completion(.success([]))
             case .empty:
                 completion(.success([]))
@@ -61,6 +60,8 @@ public final class LocalFeedLoader {
         store.retrieve { [unowned self] result in
             switch result {
             case .failure:
+                self.store.deleteCachedFeed { _ in }
+            case let .found(feed: _, timestamp) where !self.validate(timestamp):
                 self.store.deleteCachedFeed { _ in }
             default: break
             }
